User story 106 NavBar routing updated the components
Updated the nav bar components to different components so  that there is no merge conflicts while routing to different  pages
