+++ package.json
@@ -10,7 +10,7 @@
     "aws-amplify": "^4.3.16",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
